chore: moved writeFunctionConfiguration to a separate utils folder
